# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: PGFPlotsGenerator API
  description: '智能图表生成系统后端接口。除标注外均需 `Authorization: Bearer <token>`。'
  version: "1.0"
  contact:
    name: PGFPlotsGenerator
servers:
  - url: http://localhost:3000
    description: 本地默认后端地址（见 src/config.js）
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/admin/login:
    post:
      tags: [认证]
      summary: 管理员登录
      description: 校验 role=admin 的账号密码，返回 7 天有效期的 JWT。免 token。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adminAccount, password]
              properties:
                adminAccount: { type: string, example: admin123 }
                password: { type: string, example: "666666" }
      responses:
        "200":
          description: 成功，返回用户信息与 token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSuccess" }
        "400":
          description: 账号不存在/权限不足或密码错误
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      description: 需先调用验证码发送接口；验证码一次性且 10 分钟有效。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, verificationCode]
              properties:
                username: { type: string, maxLength: 10, example: zhangsan }
                email: { type: string, format: email, example: user@example.com }
                password: { type: string, description: "仅字母数字 1-8 位", example: "abc123" }
                verificationCode: { type: string, example: "123456" }
      responses:
        "200":
          description: 注册成功并自动登录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSuccess" }
        "400":
          description: 字段缺失/格式错误/用户名或邮箱已注册/验证码错误
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: user@example.com }
                password: { type: string, example: "abc123" }
      responses:
        "200":
          description: 登录成功，返回 24h JWT
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSuccess" }
        "400":
          description: 用户不存在或密码错误
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/change-password:
    post:
      tags: [认证]
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, example: "abc123" }
                newPassword: { type: string, description: "仅字母数字 1-8 位", example: "abc456" }
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
        "400":
          description: 当前密码错误/新密码格式不符/新旧相同
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/change-username:
    post:
      tags: [认证]
      summary: 修改用户名
      description: 仅字母数字下划线中文，1-10 字符。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newUsername]
              properties:
                newUsername: { type: string, example: lisi }
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
        "400":
          description: 格式错误或用户名已被占用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/change-email:
    post:
      tags: [认证]
      summary: 修改邮箱
      description: 需先向新邮箱发送验证码并校验。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newEmail, verificationCode]
              properties:
                newEmail: { type: string, format: email, example: new@example.com }
                verificationCode: { type: string, example: "123456" }
      responses:
        "200":
          description: 修改成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageSuccess" }
        "400":
          description: 格式错误/验证码错误/邮箱被占用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/auth/validate:
    get:
      tags: [认证]
      summary: 校验 token
      description: 返回当前登录用户（仅 user_id/username/email）。
      responses:
        "200":
          description: token 有效
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: 未提供或 token 无效/用户不存在
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }

  # ===================== AI 生成 /api/chat =====================
...
  /api/compile/{history_id}/pdf:
    get:
      tags: [编译]
      summary: 获取已生成 PDF 文件（鉴权 + 归属校验）
      description: 校验 generation_history 归属当前用户后 res.sendFile 流式返回 PDF，替代原 /storage 无鉴权静态托管。
      parameters:
        - name: history_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 返回 PDF 二进制流（application/pdf）
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: 历史不存在/无权访问/未生成或文件缺失
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
  /api/compile/{history_id}:
    post:
      tags: [编译]
      summary: 编译历史代码为 PDF
      description: 只读取库内 generation_code 编译（不接受 body 覆盖代码）；成功则移动到 storage/generated_charts 并更新 generation_path。
      parameters:
        - name: history_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 编译成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      history_id: { type: integer }
                      pdf_path: { type: string }
                      file_size: { type: integer }
        "400":
          description: 历史无效或无可用图表代码
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
        "404":
          description: 历史不存在或无权访问
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }
        "500":
          description: XeLaTeX 编译失败（含 stderr）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageFail" }

  # ===================== 历史 /api/history =====================
...
  /api/admin/users/{id}/reset-password:
    patch:
      tags: [管理员]
      summary: 重置用户密码为 666666
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/users/{id}:
    delete:
      tags: [管理员]
      summary: 删除用户（代码内级联删除关联数据）
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/admin/notices:
    get:
      tags: [管理员]
      summary: 系统通知列表（已排除反馈通知，含已读统计）
      security: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    post:
      tags: [管理员]
      summary: 发布系统通知（广播 target_user_id=NULL）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string, maxLength: 100 }
                content: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/notices/{id}:
    get:
      tags: [管理员]
      summary: 通知详情
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags: [管理员]
      summary: 更新通知
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    delete:
      tags: [管理员]
      summary: 删除通知
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: '用户/管理员登录返回的 JWT（请求头 `Authorization: Bearer <token>`）'
  schemas:
    User:
      type: object
      properties:
        user_id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, admin], description: 管理员登录返回中携带 }
    AuthSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        user: { $ref: "#/components/schemas/User" }
        token: { type: string, description: JWT，用户 24h / 管理员 7d 有效 }
    MessageSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
    MessageFail:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
...